Sort tree model just-in-time
